Include support for modbase structures
git-svn-id: http://chianti.ucsd.edu/svn@13074 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsf/scooter/structureViz/src/structureViz/model/ChimeraChain.java
@@ -54,7 +54,7 @@ public class ChimeraChain implements ChimeraStructuralObject {
 	/**
 	 * The model number this chain is a part of
 	 */
-	private int model;
+	private float model;
 
 	/**
 	 * A pointer to the model this chain is a part of
@@ -88,7 +88,7 @@ public class ChimeraChain implements ChimeraStructuralObject {
 	 * @param model	the model number this chain is part of
 	 * @param chainId the chain ID for this chain
 	 */
-	public ChimeraChain(int model, String chainId) {
+	public ChimeraChain(float model, String chainId) {
 		this.model = model;
 		this.chainId = chainId;
 		residueList = new TreeMap();
@@ -161,7 +161,7 @@ public class ChimeraChain implements ChimeraStructuralObject {
 	 *
 	 * @return the model number
 	 */
-	public int getModelNumber() { return model; }
+	public float getModelNumber() { return model; }
 		
 	/**
 	 * Return a string representation of this chain as follows:
